requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.4
pyyaml>=6.0
torch>=2.3
pyarrow>=15.0